More cleanup of the gurps.js file
