+++ lib/utilities.mjs
@@ -46,3 +46,19 @@ export function makeSelect(array) {
   }
   return ans;
 }
+
+// Trick to make a nice break between items, instead of "---"
+export function horiz(text, size = 10) {
+  let s = "<span style='text-decoration:line-through'>";
+  let line = s;
+  for (let i = 0; i < size; i++)
+    line += "&nbsp;";
+  line += "</span>";
+  line += " " + text + " ";
+  line += s;
+  for (let i = 0; i < size; i++)
+    line += "&nbsp;";
+  line += "</span>";
+  return line;
+}
+
